backend/requirement.txt
fastapi
uvicorn
pymongo
motor
requests
python-dotenv
